requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
python-dotenv
duckduckgo-search


# pip install -r requirements.txt  
# uvicorn main:app --reload --host 127.0.0.1 --port 8000
